* Added lps2lts to mcrl2-gui tool * Fixed compiler warning
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7841 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/utilities/include/mcrl2/utilities/mcrl2_gui_tool.h
@@ -114,7 +114,7 @@ protected:
      * \param[in] index that denotes the default value
      * \return struct of a widget_option containing the widget that represent a checkbox
      **/
-	widget_option create_radiobox_widget(std::vector<std::string> values, int index=0){
+	widget_option create_radiobox_widget(std::vector<std::string> values, size_t index=0){
 		widget_option wo;
 		if( index < values.size() ){
 			wo.default_value = values[index];
